backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai==1.3.0
opencv-python==4.8.1.78
numpy==1.26.2
python-dotenv==1.0.0
google-generativeai==0.8.5
